+++ R/last_date_boe.R
@@ -1,0 +1,17 @@
+#' Last day of BOE
+#'
+#' Returns the date of the last BOE under normal circumstances.
+#' @return A date.
+#' @export
+#' @examples
+#' last_date_boe()
+last_date_boe <- function() {
+    today <- Sys.Date()
+    time <- as.numeric(format(Sys.time(), format = "%H%I", tz = "Europe/Madrid"))
+    # On Sunday there is not BOE unless there is an extraordinary publication
+    # Also it is published at 730 local time so we check for it
+    if (weekdays(today) == "Sunday" || time < 730L) {
+        today <- today - 1
+    }
+    today
+}